// src/app/models/person.model.ts
export interface person {

    id_person?: number;
    identification: string;
    first_name: string;
    last_name: string;
    f_birthdate: string;/* Date; */
    place_birth: string;
    phone: string;
    email: string;
    address: string;
    gender: string;
    charge: string;
    ocupation: string;
    provenance: string;
    f_creation: string;/* Date; */
    f_update: string;/* Date; */
}